rename feature to `draft-pr-notice`
