+++ readme.md
@@ -373,7 +373,7 @@ Thanks for contributing! 🦋🙌
 - [](# "netiquette") [Adds unobtrusive netiquette reminders.](https://user-images.githubusercontent.com/1402241/226551766-0e1b6b15-65a3-427e-8bb5-9ea7873993be.png)
 - [](# "warn-pr-from-master") [Warns you when creating a PR from the default branch, as it’s an anti-pattern.](https://user-images.githubusercontent.com/1402241/52543516-3ca94e00-2de5-11e9-9f80-ff8f9fe8bdc4.png)
 - [](# "warning-for-disallow-edits") [Warns you when unchecking `Allow edits from maintainers`, as it’s maintainer-hostile.](https://user-images.githubusercontent.com/1402241/53151888-24101380-35ef-11e9-8d30-d6315ad97325.gif)
-- [](# "comment-on-draft-pr-indicator") [Reminds you you’re commenting on a draft PR.](https://github.com/refined-github/refined-github/assets/44045911/988a6ec1-5faa-4500-baa9-0c3713468ca9)
+- [](# "draft-pr-notice") [Reminds you you’re commenting on a draft PR.](https://github.com/refined-github/refined-github/assets/44045911/988a6ec1-5faa-4500-baa9-0c3713468ca9)
 
 <!-- Refer to style guide above. Keep this message between sections. -->
